Refator in Validator.cs - extracted rule checking function
Before the rule determining if given packages configuration is valid was
hardcoded into main function DetermineAllPackagesToInstall() so it was
not easy to see it and to modify (turn on/off) this check.  Now it is
easy to see what rules are checked in this main function and it is easy
to add/remove new ones.